refactor: refactor internal libs
