+++ libs/universal-node-xml/__stringifiers/xmlbuilder2.js
